+++ changelog.md
@@ -1,4 +1,5 @@
 # Upcoming
+* Lowered default equipment drop chance for mobs spawned from spawners
 * Fixed Respiration (oxygen_bonus attribute) not working
 * Fixed horse riding consuming hunger